changed size_t to uint for windows compatibility

+++ lib/magic/api.rb
@@ -8,7 +8,7 @@ module Magic
     attach_function :magic_close, [:pointer], :void
     attach_function :magic_file, [:pointer, :string], :pointer
     # attach_function :magic_descriptor, [:pointer, :int], :string
-    attach_function :magic_buffer, [:pointer, :pointer, :size_t], :pointer
+    attach_function :magic_buffer, [:pointer, :pointer, :uint], :pointer
     attach_function :magic_error, [:pointer], :string
     # attach_function :magic_setflags, [:pointer, :int], :int
     attach_function :magic_load, [:pointer, :string], :int
